More triangle rasteration
Pretty ok, there are some minor floating point errors on the triangle edges though. I'll leave it for now

+++ CPURendering/BoundingBox2D.cs
@@ -2,9 +2,9 @@
 
 public struct BoundingBox2D
 {
-    public float MinX;
-    public float MaxX;
-    public float MinY;
-    public float MaxY;
+    public int MinX;
+    public int MaxX;
+    public int MinY;
+    public int MaxY;
     
 }
